pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "lariska"
version = "0.1.0"
description = "Agent orchestrator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8"]

[tool.setuptools.packages.find]
where = ["."]
include = ["lariska*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
